Quadtree class added.

+++ Dodge/include/dodge/Vec3i.hpp
@@ -14,7 +14,7 @@
 namespace Dodge {
 
 
-class Vec3i;
+class Vec3f;
 
 class Vec3i {
    public:
